Moved fixtures from multiple conftest files into separate fixtures folder (#89)
